+++ src/main/java/nl/tudelft/skills/dto/view/checkpoint/ChangeCheckpointDTO.java
@@ -15,38 +15,28 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package nl.tudelft.skills.model;
+package nl.tudelft.skills.dto.view.checkpoint;
 
-import java.time.LocalDateTime;
-
-import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import nl.tudelft.skills.model.labracore.SCPerson;
 
 @Data
-@Entity
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class TaskCompletion {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+public class ChangeCheckpointDTO {
 
 	@NotNull
-	@ManyToOne
-	private SCPerson person;
+	private Long moduleId;
 
 	@NotNull
-	@ManyToOne
-	private Task task;
+	private Long prevId;
 
-	// The timestamp can be null for the already stored completions without time indication
-	@Builder.Default
-	private LocalDateTime timestamp = LocalDateTime.now();
+	@NotNull
+	private Long newId;
+
 }