+++ src/test/java/nl/tudelft/skills/service/SkillServiceTest.java
@@ -18,6 +18,7 @@
 package nl.tudelft.skills.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.*;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -25,18 +26,20 @@ import java.util.stream.Collectors;
 
 import nl.tudelft.labracore.api.CourseControllerApi;
 import nl.tudelft.labracore.api.EditionControllerApi;
-import nl.tudelft.labracore.api.dto.CourseDetailsDTO;
-import nl.tudelft.labracore.api.dto.EditionDetailsDTO;
-import nl.tudelft.labracore.api.dto.EditionSummaryDTO;
+import nl.tudelft.labracore.api.RoleControllerApi;
+import nl.tudelft.labracore.api.dto.*;
 import nl.tudelft.skills.TestSkillCircuitsApplication;
 import nl.tudelft.skills.model.*;
 import nl.tudelft.skills.repository.*;
+import nl.tudelft.skills.security.AuthorisationService;
 import nl.tudelft.skills.test.TestDatabaseLoader;
+import nl.tudelft.skills.test.TestUserDetailsService;
 
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.test.context.support.WithUserDetails;
 import org.springframework.transaction.annotation.Transactional;
 
 import reactor.core.publisher.Flux;
@@ -54,7 +57,9 @@ public class SkillServiceTest {
 	private final ModuleRepository moduleRepository;
 	private final SkillService skillService;
 	private final SkillRepository skillRepository;
+	private AuthorisationService authorisationService;
 
+	private final RoleControllerApi roleApi;
 	private final CourseControllerApi courseApi;
 	private final EditionControllerApi editionApi;
 
@@ -66,7 +71,8 @@ public class SkillServiceTest {
 			TestDatabaseLoader db, TaskCompletionRepository taskCompletionRepository,
 			EditionControllerApi editionApi, CourseControllerApi courseApi,
 			SkillRepository skillRepository, ModuleRepository moduleRepository,
-			ExternalSkillRepository externalSkillRepository, EditionRepository editionRepository) {
+			ExternalSkillRepository externalSkillRepository, EditionRepository editionRepository,
+			AuthorisationService authorisationService, RoleControllerApi roleApi) {
 		this.abstractSkillRepository = abstractSkillRepository;
 		this.taskCompletionRepository = taskCompletionRepository;
 		this.externalSkillRepository = externalSkillRepository;
@@ -75,14 +81,19 @@ public class SkillServiceTest {
 		this.taskRepository = taskRepository;
 		this.skillRepository = skillRepository;
 
+		// The service is not mocked to test the specifics of whether an edition is shown because it
+		// is visible, or because the person is at least a teacher in the edition
+		this.authorisationService = authorisationService;
+
 		this.db = db;
 		this.localDateTime = LocalDateTime.of(2023, 1, 10, 10, 10, 0);
 
+		this.roleApi = roleApi;
 		this.courseApi = courseApi;
 		this.editionApi = editionApi;
 
 		this.skillService = new SkillService(abstractSkillRepository, taskCompletionRepository, editionApi,
-				courseApi, skillRepository, editionRepository);
+				courseApi, skillRepository, editionRepository, authorisationService);
 	}
 
 	@Test
@@ -161,30 +172,13 @@ public class SkillServiceTest {
 	}
 
 	/**
-	 * Creates an external skill for testing purposes.
-	 *
-	 * @return The created external skill.
-	 */
-	private ExternalSkill createExternalSkill() {
-		// Create a new module for the external skill, in the same edition
-		SCModule module = moduleRepository.save(SCModule.builder().edition(db.getEditionRL())
-				.name("New module").build());
-		// Create an external skill referencing SkillAssumption
-		ExternalSkill externalSkill = ExternalSkill.builder().skill(db.getSkillAssumption()).module(module)
-				.row(0).column(0).build();
-		externalSkill = externalSkillRepository.save(externalSkill);
-
-		return externalSkill;
-	}
-
-	/**
 	 * Mocks the responses of the courseApi and editionApi for three courses. Sets visibility of editionRL to
-	 * the given value, and mocks editions of given ids to be active.
+	 * the given value. Also mocks the role of the person in all editions.
 	 *
-	 * @param activeEditionIds The ids of active editions.
-	 * @param visible          Whether the current edition should be visible.
+	 * @param visible Whether the current edition should be visible.
+	 * @param role    Role of the person.
 	 */
-	private void mockEditionsAndSetVisible(List<Long> activeEditionIds, boolean visible) {
+	private void mockEditionsAndSetVisible(boolean visible, String role) {
 		Long idInUse = db.getEditionRL().getId();
 
 		// Mock the response of the courseApi to return the course details
@@ -196,24 +190,29 @@ public class SkillServiceTest {
 						new EditionSummaryDTO().id(idInUse + 2).startDate(localDateTime.plusDays(4))));
 		Mockito.when(courseApi.getCourseByEdition(db.getEditionRL().getId())).thenReturn(Mono.just(course));
 
-		// Mock response so that the editions are all also active
-		List<EditionDetailsDTO> activeEditions = activeEditionIds.stream()
-				.map(id -> new EditionDetailsDTO().id(id))
-				.collect(Collectors.toList());
-		Mockito.when(editionApi.getAllEditionsActiveOrTaughtBy(db.getPerson().getId()))
-				.thenReturn(Flux.fromIterable(activeEditions));
-
 		// Set edition to be visible
 		db.getEditionRL().setVisible(visible);
 		editionRepository.save(db.getEditionRL());
+
+		// Mock the role of the person for all of the editions
+		List<RoleDetailsDTO> roleDetails = List.of(idInUse, idInUse + 1, idInUse + 2).stream()
+				.map(id -> new RoleDetailsDTO()
+						.id(new Id().editionId(id)
+								.personId(TestUserDetailsService.id))
+						.person(new PersonSummaryDTO().id(TestUserDetailsService.id).username("username"))
+						.type(RoleDetailsDTO.TypeEnum.valueOf(role)))
+				.collect(Collectors.toList());
+		when(roleApi.getRolesById(anySet(), anySet()))
+				.thenReturn(Flux.fromIterable(roleDetails));
 	}
 
 	@Test
+	@WithUserDetails("username")
 	public void testRecentActiveEditionOneEdition() {
 		// Test scenario in which there is only one edition
 		// The method should return that editions skill
 
-		ExternalSkill externalSkill = createExternalSkill();
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
 
 		// Mock the response of the courseApi to return the course details
 		CourseDetailsDTO course = new CourseDetailsDTO().id(db.getCourseRL().getId())
@@ -227,6 +226,13 @@ public class SkillServiceTest {
 		// Set edition to be visible
 		db.getEditionRL().setVisible(true);
 		editionRepository.save(db.getEditionRL());
+		// Mock the role of the person
+		when(roleApi.getRolesById(anySet(), anySet()))
+				.thenReturn(Flux.just(new RoleDetailsDTO()
+						.id(new Id().editionId(db.getEditionRL().getId())
+								.personId(TestUserDetailsService.id))
+						.person(new PersonSummaryDTO().id(TestUserDetailsService.id).username("username"))
+						.type(RoleDetailsDTO.TypeEnum.valueOf("STUDENT"))));
 
 		// Assert that the recent active edition method returns the correct skill
 		// Since there is only one edition, this is the skill in this edition
@@ -239,6 +245,39 @@ public class SkillServiceTest {
 	}
 
 	@Test
+	@WithUserDetails("username")
+	public void testRecentActiveEditionReturnNull() {
+		// Test scenario in which there is only a skill in an invisible edition
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
+
+		// Mock the response of the courseApi to return the course details
+		CourseDetailsDTO course = new CourseDetailsDTO().id(db.getCourseRL().getId())
+				.editions(List.of(new EditionSummaryDTO().id(db.getEditionRL().getId())));
+		Mockito.when(courseApi.getCourseByEdition(db.getEditionRL().getId())).thenReturn(Mono.just(course));
+
+		// Mock response so that the edition is also active
+		Mockito.when(editionApi.getAllEditionsActiveOrTaughtBy(db.getPerson().getId()))
+				.thenReturn(Flux.just(new EditionDetailsDTO().id(db.getEditionRL().getId())));
+
+		// Mock the role of the person
+		when(roleApi.getRolesById(anySet(), anySet()))
+				.thenReturn(Flux.just(new RoleDetailsDTO()
+						.id(new Id().editionId(db.getEditionRL().getId())
+								.personId(TestUserDetailsService.id))
+						.person(new PersonSummaryDTO().id(TestUserDetailsService.id).username("username"))
+						.type(RoleDetailsDTO.TypeEnum.valueOf("STUDENT"))));
+
+		// Assert that the recent active edition method returns null (edition is invisible)
+		assertThat(skillService.recentActiveEditionForSkillOrLatest(db.getPerson().getId(), externalSkill))
+				.isEqualTo(null);
+
+		// Assert on the traversal list
+		assertThat(skillService.traverseSkillTree(db.getSkillAssumption()))
+				.containsExactly(db.getSkillAssumption());
+	}
+
+	@Test
+	@WithUserDetails("username")
 	public void testMultipleEditionsNoTaskCompleted() {
 		/*
 		 * Test scenario in which there are multiple editions, but no task was completed. Edition structure
@@ -247,7 +286,7 @@ public class SkillServiceTest {
 		 * should return the latest editions skill (edition C).
 		 */
 
-		ExternalSkill externalSkill = createExternalSkill();
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
 
 		// Reset the task completions, so that the person has not completed any tasks yet
 		db.resetTaskCompletions();
@@ -264,7 +303,7 @@ public class SkillServiceTest {
 		db.getSkillAssumption().getFutureEditionSkills().add(skillEditionB);
 		db.getSkillAssumption().getFutureEditionSkills().add(skillEditionC);
 
-		mockEditionsAndSetVisible(List.of(idInUse, idInUse + 1, idInUse + 2), true);
+		mockEditionsAndSetVisible(true, "STUDENT");
 
 		// Assert that the recent active edition method returns the correct skill
 		// The person has not completed any tasks in any edition yet, so it should return the most recent
@@ -283,6 +322,7 @@ public class SkillServiceTest {
 	}
 
 	@Test
+	@WithUserDetails("username")
 	public void testMultipleEditionsWithTaskCompleted() {
 		/*
 		 * Test scenario in which there are multiple editions, and a task was completed in one of the less
@@ -292,7 +332,7 @@ public class SkillServiceTest {
 		 * in which a task was completed.
 		 */
 
-		ExternalSkill externalSkill = createExternalSkill();
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
 
 		// Reset the task completions, so that the person has not completed any tasks yet
 		db.resetTaskCompletions();
@@ -317,7 +357,7 @@ public class SkillServiceTest {
 		task.getCompletedBy().add(taskCompletion);
 		db.getPerson().getTaskCompletions().add(taskCompletion);
 
-		mockEditionsAndSetVisible(List.of(idInUse, idInUse + 1, idInUse + 2), true);
+		mockEditionsAndSetVisible(true, "STUDENT");
 
 		// Assert that the recent active edition method returns the correct skill
 		// The person has completed a task in editionB, so it should return that editions skill
@@ -335,6 +375,7 @@ public class SkillServiceTest {
 	}
 
 	@Test
+	@WithUserDetails("username")
 	public void testMultipleEditionsChainWithTaskCompleted() {
 		/*
 		 * Test scenario in which there are multiple editions, and a task was completed in one of the less
@@ -344,7 +385,7 @@ public class SkillServiceTest {
 		 * in which a task was completed.
 		 */
 
-		ExternalSkill externalSkill = createExternalSkill();
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
 
 		// Reset the task completions, so that the person has not completed any tasks yet
 		db.resetTaskCompletions();
@@ -370,7 +411,7 @@ public class SkillServiceTest {
 		task.getCompletedBy().add(taskCompletion);
 		db.getPerson().getTaskCompletions().add(taskCompletion);
 
-		mockEditionsAndSetVisible(List.of(idInUse, idInUse + 1, idInUse + 2), true);
+		mockEditionsAndSetVisible(true, "STUDENT");
 
 		// Assert that the recent active edition method returns the correct skill
 		// The person has completed a task in editionB, so it should return that editions skill
@@ -387,6 +428,7 @@ public class SkillServiceTest {
 	}
 
 	@Test
+	@WithUserDetails("username")
 	public void testMultipleEditionsMostRecentDoesNotHaveSkill() {
 		/*
 		 * Test scenario in which there are multiple editions, and the most recent edition does not contain
@@ -396,7 +438,7 @@ public class SkillServiceTest {
 		 * return the skill in edition B.
 		 */
 
-		ExternalSkill externalSkill = createExternalSkill();
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
 
 		// Reset the task completions, so that the person has not completed any tasks yet
 		db.resetTaskCompletions();
@@ -413,7 +455,7 @@ public class SkillServiceTest {
 		// Create empty edition C
 		editionRepository.save(SCEdition.builder().id(idInUse + 2).build());
 
-		mockEditionsAndSetVisible(List.of(idInUse, idInUse + 1, idInUse + 2), true);
+		mockEditionsAndSetVisible(true, "STUDENT");
 
 		// Assert that the recent active edition method returns the correct skill
 		// The most recent edition which contains the copied skill is editionB
@@ -428,6 +470,7 @@ public class SkillServiceTest {
 	}
 
 	@Test
+	@WithUserDetails("username")
 	public void testMultipleEditionsTaskCompletedInvisibleEdition() {
 		/*
 		 * Test scenario in which there are multiple editions, and a task was completed in the oldest edition,
@@ -437,7 +480,7 @@ public class SkillServiceTest {
 		 * most skill of recent edition, which is also visible (editionC).
 		 */
 
-		ExternalSkill externalSkill = createExternalSkill();
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
 
 		// Reset the task completions, so that the person has not completed any tasks yet
 		db.resetTaskCompletions();
@@ -462,7 +505,7 @@ public class SkillServiceTest {
 		task.getCompletedBy().add(taskCompletion);
 		db.getPerson().getTaskCompletions().add(taskCompletion);
 
-		mockEditionsAndSetVisible(List.of(idInUse, idInUse + 1, idInUse + 2), false);
+		mockEditionsAndSetVisible(false, "STUDENT");
 
 		// Assert that the recent active edition method returns the correct skill
 		// The latest visible edition is edition C, so it should return the skill in that edition
@@ -480,6 +523,7 @@ public class SkillServiceTest {
 	}
 
 	@Test
+	@WithUserDetails("username")
 	public void testMultipleEditionsWithoutTaskCompletedInvisibleEdition() {
 		/*
 		 * Test scenario in which there are multiple editions, and the most recent edition is not visible.
@@ -488,7 +532,7 @@ public class SkillServiceTest {
 		 * The method should return the most skill of recent edition, which is also visible (editionB).
 		 */
 
-		ExternalSkill externalSkill = createExternalSkill();
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
 
 		// Reset the task completions, so that the person has not completed any tasks yet
 		db.resetTaskCompletions();
@@ -505,12 +549,59 @@ public class SkillServiceTest {
 		db.getSkillAssumption().getFutureEditionSkills().add(skillEditionB);
 		db.getSkillAssumption().getFutureEditionSkills().add(skillEditionC);
 
-		mockEditionsAndSetVisible(List.of(idInUse, idInUse + 1, idInUse + 2), true);
+		mockEditionsAndSetVisible(true, "STUDENT");
 
 		// Assert that the recent active edition method returns the correct skill
 		// The latest visible edition is edition B, so it should return the skill in that edition
 		assertThat(skillService.recentActiveEditionForSkillOrLatest(db.getPerson().getId(), externalSkill))
 				.isEqualTo(skillEditionB);
+
+		// Assert on the traversal lists
+		// The order cannot be asserted on, since the futureEditionSkills is a set
+		assertThat(skillService.traverseSkillTree(db.getSkillAssumption()))
+				.containsExactlyInAnyOrder(db.getSkillAssumption(), skillEditionB, skillEditionC);
+		assertThat(skillService.traverseSkillTree(skillEditionB))
+				.containsExactlyInAnyOrder(db.getSkillAssumption(), skillEditionB, skillEditionC);
+		assertThat(skillService.traverseSkillTree(skillEditionC))
+				.containsExactlyInAnyOrder(db.getSkillAssumption(), skillEditionB, skillEditionC);
+	}
+
+	@Test
+	@WithUserDetails("teacher")
+	public void testMultipleEditionsWithoutTaskCompletedTeacherEdition() {
+		/*
+		 * Test scenario in which there are multiple editions, and the most recent edition is not visible, the
+		 * logged-in user is however a teacher in these editions, so it should be visible to them regardless.
+		 * Edition structure is: editionA --> (editionB and editionC). With editionC being the latest edition.
+		 * The skill which the external skill refers to should be "copied" from edition A to editions B and C.
+		 * The method should return the most skill of recent edition, which is also visible to the user
+		 * (editionC).
+		 */
+
+		ExternalSkill externalSkill = db.createExternalSkill(db.getSkillAssumption());
+
+		// Reset the task completions, so that the person has not completed any tasks yet
+		db.resetTaskCompletions();
+
+		// Create two new skills, submodules, modules and editions
+		// Edition A is the current edition (editionRL)
+		Long idInUse = db.getEditionRL().getId();
+		Skill skillEditionB = db.createSkillInEditionHelper(idInUse + 1, true);
+		Skill skillEditionC = db.createSkillInEditionHelper(idInUse + 2, false);
+
+		// Skills in edition B/C should be copies of skill in edition A
+		skillEditionB.setPreviousEditionSkill(db.getSkillAssumption());
+		skillEditionC.setPreviousEditionSkill(db.getSkillAssumption());
+		db.getSkillAssumption().getFutureEditionSkills().add(skillEditionB);
+		db.getSkillAssumption().getFutureEditionSkills().add(skillEditionC);
+
+		// Sets the role to "teacher" for all editions
+		mockEditionsAndSetVisible(true, "TEACHER");
+
+		// Assert that the recent active edition method returns the correct skill
+		// The latest visible edition is edition B, so it should return the skill in that edition
+		assertThat(skillService.recentActiveEditionForSkillOrLatest(db.getPerson().getId(), externalSkill))
+				.isEqualTo(skillEditionC);
 
 		// Assert on the traversal lists
 		// The order cannot be asserted on, since the futureEditionSkills is a set